requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.5.2
async-timeout==4.0.3
attrs==25.3.0
blinker==1.8.2
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.3.0
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib-resources==6.4.5
jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langchain==0.2.17
langchain-community==0.2.19
langchain-core==0.2.43
langchain-openai==0.1.25
langchain-text-splitters==0.2.4
langsmith==0.1.147
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.1.0
narwhals==1.39.1
numpy==1.24.4
openai==1.79.0
orjson==3.10.15
packaging==24.2
pandas==2.0.3
pillow==10.4.0
pkgutil-resolve-name==1.3.10
propcache==0.2.0
protobuf==5.29.4
pyarrow==17.0.0
pydantic==2.10.6
pydantic-core==2.27.2
pydeck==0.9.1
pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlalchemy==2.0.41
streamlit==1.40.1
tenacity==8.5.0
tiktoken==0.7.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing-extensions==4.13.2
typing-inspect==0.9.0
tzdata==2025.2
urllib3==2.2.3
yarl==1.15.2
zipp==3.20.2
